requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cli-helpers==2.1.0
click==7.1.2
configobj==5.0.6
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-ckeditor==6.0.0
django-colorfield==0.4.1
django-crispy-forms==1.9.1
django-currentuser==0.5.0
django-js-asset==1.2.2
django-nested-inline==0.4.2
django-rosetta==0.9.4
django-translations==1.1.2
humanize==3.4.1
idna==2.10
oauthlib==3.1.0
pgspecial==1.11.5
Pillow==8.1.0
polib==1.1.0
postgres==3.0.0
prompt-toolkit==1.0.18
psycopg2-binary==2.8.6
psycopg2-pool==1.1
Pygments==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
setproctitle==1.2.2
six==1.15.0
sqlparse==0.4.1
tabulate==0.8.9
terminaltables==3.1.0
urllib3==1.26.3
wcwidth==0.2.5
